feat: initial team support, RBAC, avatars

+++ app/Http/Controllers/Api/AccessTokenController.php
@@ -10,16 +10,21 @@ use App\Values\AccessToken;
 use Bag\Attributes\WithoutValidation;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class AccessTokenController extends Controller
 {
     public function store(AccessToken $accessToken, AccessTokenService $accessTokenService): JsonResponse
     {
+        Gate::authorize('update', $accessToken);
+
         return response()->json($accessTokenService->create($accessToken));
     }
 
     public function show(AccessTokenService $accessTokenService): JsonResponse
     {
+        Gate::authorize('viewAny', AccessToken::class);
+
         return response()->json($accessTokenService->all());
     }
 
@@ -28,6 +33,8 @@ class AccessTokenController extends Controller
         AccessToken $accessToken,
         AccessTokenService $accessTokenService
     ): Response {
+        Gate::authorize('delete', $accessToken);
+
         if ($accessTokenService->delete($accessToken)) {
             return response(null, 204);
         }
